Move log to general lib
