+++ ZelBack/src/lib/server.js
@@ -1,8 +1,6 @@
 const express = require('express');
 const eWS = require('express-ws');
 const cors = require('cors');
-
-const log = require('./log');
 const morgan = require('morgan');
 const compression = require('compression');
 
@@ -11,14 +9,14 @@ const log = require('./log');
 const expressWs = eWS(express());
 const { app } = expressWs;
 
-const logger = () => {
-  return (req, res, next) => {
-    log.debug({ url: req.url, method: req.method, ip: req.ip.replace('::ffff:', '') }, "Incomming API request:")
-    next();
-  }
+const logger = (req, res, next) => {
+  log.info(`\n\n${req.method}\n${req.url}\n${req.ip}\n`)
+  next();
 };
 
-app.use(logger());
+app.use(compression());
+app.use(morgan('combined'));
+app.use(logger);
 app.use(express.json());
 // app.use(express.urlencoded({ extended: true }));
 app.use(cors());
