+++ ZelBack/src/lib/socket.js
@@ -3,7 +3,7 @@ const verificationHelperUtils = require('../services/verificationHelperUtils');
 const dockerService = require('../services/dockerService');
 const serviceHelper = require('../services/serviceHelper');
 
-const log = require('./log');
+const log = require('../../../lib/log');
 
 let io;
