+++ ZelBack/src/services/backupRestoreService.js
@@ -1,4 +1,4 @@
-const log = require('../lib/log');
+const log = require('../../../../lib/log');
 const path = require('path');
 const messageHelper = require('./messageHelper');
 const verificationHelper = require('./verificationHelper');
