+++ ZelBack/src/services/fluxCommunication.js
@@ -3,7 +3,7 @@ const config = require('config');
 const { LRUCache } = require('lru-cache');
 const hash = require('object-hash');
 const WebSocket = require('ws');
-const log = require('../lib/log');
+const log = require('../../../../lib/log');
 const serviceHelper = require('./serviceHelper');
 const verificationHelper = require('./verificationHelper');
 const daemonServiceMiscRpcs = require('./daemonService/daemonServiceMiscRpcs');
