+++ ZelBack/src/services/idService.js
@@ -1,7 +1,7 @@
 const config = require('config');
 const qs = require('qs');
 const os = require('os');
-const log = require('../lib/log');
+const log = require('../../../../lib/log');
 const serviceHelper = require('./serviceHelper');
 const messageHelper = require('./messageHelper');
 const dbHelper = require('./dbHelper');
